test(proxy): split nested-access-groups test file under size gate
577-LOC test file tripped the +500 added-LOC size gate. classify +
validate coverage moves to a sibling file; the original keeps the DFS
and edge-case suite. each file is now under 500 LOC.
